Correcting the object dimension deletion unit tests.
For one thing, tuples are immutable, so we can't use one to test for
successful deletions.

Furthermore, deleting the second item in the list of three items will
result in a list with two items, and the second slot with this contain a
valid item: the item previously in the third slot.